Fix countsToFrequencies using decimal64 precision (#17)
FrequencyUtils.countsToFrequencies is dividing every count by the total, but double precision generates frequencies that does not sum up to one. This is an issue while validating the frequencies with our framework, and it is solved in this commit by using BigDecimal division with decimal64 precision.

In addition, this commit simplify the code in NucleotideDiversity to use this method.